Add: Firebase 연동

+++ app/src/main/java/com/example/noogabab/presentation/entity/PresenterAlbumImage.kt
@@ -2,4 +2,4 @@ package com.example.noogabab.presentation.entity
 
 import java.io.Serializable
 
-data class PresenterAlbumImage(val id: Int, val image: String): Serializable
+data class PresenterAlbumImage(val id: Long, val image: String): Serializable
